Add steering evaluation results and fix bfloat16 SVD bug
Steering results from Modal runs (2026-04-12):
- ICV steering vector (seed 12345, layer 18, PCA, normalized)
- Validation ladder: 8 settings on medium_stakes_validation (partial, 92-184/200)
- Held-out evals: high_stakes_test, astro_deploy, steals_test (partial, 62-80/1000)
- MMLU-Redux capability retention: 69.8% (vs 71.18% base)
- Locked setting: layer 12, alpha 1.0

Fix bfloat16 SVD bug in generate_steering_vector.py: cast stacked
diffs to float32 before PCA aggregation (CUDA gesvdj doesn't
support bfloat16).

+++ generate_steering_vector.py
@@ -408,7 +408,8 @@ def main():
         sys.exit(1)
 
     # --- Aggregate to get final steering vector ---
-    stacked = torch.stack(filtered_diffs)
+    # Cast to float32 upfront so SVD and dot products work on CUDA
+    stacked = torch.stack(filtered_diffs).float()
     pca_singular_values = None
     if args.icv_method == "pca":
         centered = stacked - stacked.mean(dim=0)
